+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from './types';
 export * from './config-types';
+export * from './errors';
 export * from './runtime-model';
 export * from './identity';